// src/platform/generated/ProjectDialogAction.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ProjectDialogAction = "cancelAlbumInformation" | "cancelExport" | "cancelProjectClose" | "discardAndClose" | "confirmAlbumInformation" | "closeProjectAfterGraphicsFailure" | "dismissExport" | "dismissProjectCloseFailure" | "dismissProjectOperationFailure" | "retryExport" | "saveAndClose";